+++ controllers/contacts/updateContactById.js
@@ -1,21 +1,16 @@
 const contactsOperations = require("../../models/contacts");
-const Joi = require("joi");
+// const Joi = require("joi");
 
-const putContactSchema = Joi.object({
-  name: Joi.string().alphanum().min(3).max(30).optional(),
-  email: Joi.string().email().optional(),
-  phone: Joi.string()
-    .regex(/^[0-9]{10,15}$/)
-    .optional(),
-});
+// const putContactSchema = Joi.object({
+//   name: Joi.string().alphanum().min(3).max(30).optional(),
+//   email: Joi.string().email().optional(),
+//   phone: Joi.string()
+//     .regex(/^[0-9]{10,15}$/)
+//     .optional(),
+// });
 
 const updateContactById = async (req, res, next) => {
   try {
-    const { error } = putContactSchema.validate(req.body);
-    if (error) {
-      res.status(400).json(` Missing  fields`);
-      throw error;
-    }
     const { contactId } = req.params;
     const result = await contactsOperations.updateContactById(
       contactId,
