+++ middlewares/contactValidation.js
@@ -1,0 +1,13 @@
+const contactValidation = (schema) => {
+  return (req, res, next) => {
+    const { error } = schema.validate(req.body);
+    if (error) {
+      error.status = 400;
+      next(error);
+      return;
+    }
+    next();
+  };
+};
+
+module.exports = contactValidation;
