+++ middlewares/index.js
@@ -1,0 +1,5 @@
+const contactValidation = require("./contactValidation");
+
+module.exports = {
+  contactValidation,
+};
