+++ routes/api/contacts.js
@@ -1,5 +1,9 @@
 const express = require("express");
 
+const { contactValidation } = require("../../middlewares");
+
+const { postContactSchema, putContactSchema } = require("../../schemas");
+
 const { contacts: ctrl } = require("../../controllers");
 
 const router = express.Router();
@@ -8,9 +12,13 @@ router.get("/", ctrl.listContacts);
 
 router.get("/:contactId", ctrl.getContactById);
 
-router.post("/", ctrl.addContact);
+router.post("/", contactValidation(postContactSchema), ctrl.addContact);
 
-router.put("/:contactId", ctrl.updateContactById);
+router.put(
+  "/:contactId",
+  contactValidation(putContactSchema),
+  ctrl.updateContactById
+);
 
 router.delete("/:contactId", ctrl.removeContact);
