+++ schemas/index.js
@@ -1,0 +1,7 @@
+const postContactSchema = require("./postContactSchema");
+const putContactSchema = require("./putContactSchema");
+
+module.exports = {
+  postContactSchema,
+  putContactSchema,
+};
