+++ schemas/postContactSchema.js
@@ -1,0 +1,11 @@
+const Joi = require("joi");
+
+const postContactSchema = Joi.object({
+  name: Joi.string().alphanum().min(3).max(30).required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string()
+    .regex(/\(?([0-9]{3})\)?([ .-]?)([0-9]{3})\2([0-9]{4})/)
+    .required(),
+});
+
+module.exports = postContactSchema;
